modify hooks
