+++ src/components/organisms/setting-form/hooks.ts
@@ -1,0 +1,32 @@
+import { useDispatch } from 'react-redux';
+import {
+  updateGameStartFlg,
+  initializeBoard,
+  updateCurrentPlayer,
+  countScore,
+} from 'src/redux/modules/game';
+import Const from 'src/const';
+
+const { Player, PieceColor } = Const;
+
+const useSettingForm = (): {
+  startGame: (sideSquaresCount: number) => void;
+} => {
+  const dispatch = useDispatch();
+
+  const startGame = (sideSquaresCount: number) => {
+    dispatch(updateGameStartFlg(true));
+    dispatch(initializeBoard(sideSquaresCount));
+    dispatch(
+      updateCurrentPlayer({
+        player: Player.PLAYER_2,
+        pieceColor: PieceColor.BLACK,
+      }),
+    );
+    dispatch(countScore());
+  };
+
+  return { startGame };
+};
+
+export default useSettingForm;
